rename local_workflow to integration
